requirements.txt
defusedxml
cssselect2
pillow
img2pdf
asn1
